requirements.txt
anvil-parser>=0.9.0
numpy>=1.24.0
